+++ CaloriesCalculator/Models/DailyStats.cs
@@ -11,9 +11,9 @@ namespace CaloriesCalculator.Models
         public int ID { get; set; }
         public string Date { get; set; }
         public int Water { get; set; }
-        public int Calories { get; set; }
-        public int Proteins { get; set; }
-        public int Fats { get; set; }
-        public int Carbohydrates { get; set; }
+        public double Calories { get; set; }
+        public double Proteins { get; set; }
+        public double Fats { get; set; }
+        public double Carbohydrates { get; set; }
     }
 }
